fix(ci): line-limit gate — move corrupt-gitfile reclaim test out of test_git_manager
The missing-HEAD reclaim coverage landed in test_git_manager.py and pushed it
past the 1500-line maintainability gate; keep the behavior test beside the
ensure_worktree usability suite instead.
